index exports oneShotRegression
