+++ .jscsrc
@@ -1,0 +1,8 @@
+{
+    "esnext": true,
+    "preset": "airbnb",
+    "validateIndentation": 4,
+    "safeContextKeyword": ["self"],
+    "disallowMultipleVarDecl": "exceptUndefined",
+    "requireTrailingComma": null
+}
