+++ lib/one-shot-regression.js
@@ -45,7 +45,7 @@ module.exports = {
      * @param  {array} W set of minimized regressors
      * @return {array}
      */
-    aggregrate: function(W) {
+    aggregrateAvg: function(W) {
         var result = n.dot(
             n.rep([W.length], 1),
             W
